+++ R/reexport-xldiff.R
@@ -1,0 +1,3 @@
+#' @importFrom xldiff cell_range_translate
+#' @export
+xldiff::cell_range_translate