+++ NEWS.md
@@ -3,4 +3,9 @@
 - `row_finder()`: finds row number based on an "anchor" cell of known column and pattern.
 - `as_numeric_smart()`: as.numeric(), but correctly handles cases of numbers with commas in them, numbers ending in %s.
 - `copy_columns()`: copies specified sets of columns from start_row to end_row from one file into another, with highlighting of cells copied and cells with changed values
+- `implant_df()` and `copy_section()` to copy sections of excel sheets
+- addition of `clip_to_vec()` to make coding easier
 - addition of `validate_*()` for input validation
+
+TODO: 
+ - confirm fram r-universe is including this